AP-279 refactor namespace PaymentMethodWriter
