+++ Resources/services/dbal.xml
@@ -17,7 +17,7 @@
         </service>
         <service
                 id="adyen_payment.dbal.payment_attributes"
-                class="AdyenPayment\Dbal\PaymentAttributes"
+                class="AdyenPayment\Dbal\Provider\Payment\Attributes\PaymentAttributeProvider"
         >
             <argument type="service" id="db"/>
         </service>
